requirements.txt
pydantic
numpy
pillow
httpx
fastapi[standard]

# the following installs the package we are currently developing
.
